requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
requests>=2.31.0
google-cloud-storage>=3.1.1
google-auth>=2.25.0
protobuf>=4.25.0
